Move websocket query bridge into exporter v1 API
