+++ exporter/v1/api/query_bridge_test.go
@@ -1,4 +1,4 @@
-package operator
+package api
 
 import (
 	"encoding/hex"
@@ -17,10 +17,8 @@ import (
 	dutytracer "github.com/ssvlabs/ssv/exporter/dutytracer"
 	dutytracestore "github.com/ssvlabs/ssv/exporter/store"
 	"github.com/ssvlabs/ssv/exporter/traces"
-	"github.com/ssvlabs/ssv/exporter/v1/api"
 	ibftstorage "github.com/ssvlabs/ssv/ibft/storage"
 	"github.com/ssvlabs/ssv/networkconfig"
-	"github.com/ssvlabs/ssv/operator/validator"
 	registrystoragemocks "github.com/ssvlabs/ssv/registry/storage/mocks"
 	kv "github.com/ssvlabs/ssv/storage/badger"
 	"github.com/ssvlabs/ssv/storage/basedb"
@@ -61,7 +59,8 @@ func (s *faultingDutyTraceStore) GetValidatorDuty(slot phase0.Slot, role spectyp
 type wsQueryHarness struct {
 	t *testing.T
 
-	node *Node
+	handler      *Handler
+	exporterRead *exportercore.Exporter
 
 	ctrl          *gomock.Controller
 	validatorMock *registrystoragemocks.MockValidatorStore
@@ -87,18 +86,10 @@ func newWSQueryHarness(t *testing.T) *wsQueryHarness {
 	collector := dutytracer.New(zap.NewNop(), validatorMock, nil, store, networkconfig.TestNetwork.Beacon, nil, nil)
 	coreExporter := exportercore.NewExporter(zap.NewNop(), ibftstorage.NewStores(), collector, validatorMock)
 
-	node := &Node{
-		logger:       logger,
-		network:      networkconfig.TestNetwork,
-		exporterRead: coreExporter,
-		validatorOptions: validator.ControllerOptions{
-			ValidatorStore: validatorMock,
-		},
-	}
-
 	return &wsQueryHarness{
 		t:             t,
-		node:          node,
+		handler:       NewHandler(logger),
+		exporterRead:  coreExporter,
 		ctrl:          ctrl,
 		validatorMock: validatorMock,
 		db:            db,
@@ -155,17 +146,17 @@ func (h *wsQueryHarness) SaveCommitteeDutyAttester(slot phase0.Slot, validatorIn
 	require.NoError(h.t, h.store.SaveCommitteeDuty(duty))
 }
 
-func (h *wsQueryHarness) QueryDecided(from, to uint64, pkHex, role string) *api.NetworkMessage {
+func (h *wsQueryHarness) QueryDecided(from, to uint64, pkHex, role string) *NetworkMessage {
 	h.t.Helper()
 	nm := decidedMessage(from, to, pkHex, role)
-	h.node.handleQueryRequests(nm)
+	h.handler.HandleQueryRequests(nil, h.exporterRead, h.validatorMock, networkconfig.TestNetwork.DomainType, nm)
 	return nm
 }
 
-func decidedMessage(from, to uint64, pkHex, role string) *api.NetworkMessage {
-	return &api.NetworkMessage{Msg: api.Message{
-		Type: api.TypeDecided,
-		Filter: api.MessageFilter{
+func decidedMessage(from, to uint64, pkHex, role string) *NetworkMessage {
+	return &NetworkMessage{Msg: Message{
+		Type: TypeDecided,
+		Filter: MessageFilter{
 			From:      from,
 			To:        to,
 			PublicKey: pkHex,
@@ -189,9 +180,9 @@ func TestWSQuery_Decided_Proposer_SingleSlot(t *testing.T) {
 	h.SaveValidatorDuty(phase0.Slot(10), spectypes.BNRoleProposer, idx, []spectypes.OperatorID{11, 12})
 
 	nm := h.QueryDecided(uint64(10), uint64(10), hex.EncodeToString(pk[:]), spectypes.BNRoleProposer.String())
-	require.Equal(t, api.TypeDecided, nm.Msg.Type)
+	require.Equal(t, TypeDecided, nm.Msg.Type)
 	require.Equal(t, nm.Msg.Filter.PublicKey, hex.EncodeToString(pk[:]))
-	data, ok := nm.Msg.Data.([]*api.ParticipantsAPI)
+	data, ok := nm.Msg.Data.([]*ParticipantsAPI)
 	require.True(t, ok, "response data should be participants slice")
 	require.Len(t, data, 1)
 	require.Equal(t, uint64(10), uint64(data[0].Slot))
@@ -211,8 +202,8 @@ func TestWSQuery_Decided_Attester_UsesCommitteeLookup(t *testing.T) {
 
 	nm := h.QueryDecided(3, 3, hex.EncodeToString(pk[:]), spectypes.BNRoleAttester.String())
 
-	require.Equal(t, api.TypeDecided, nm.Msg.Type)
-	data, ok := nm.Msg.Data.([]*api.ParticipantsAPI)
+	require.Equal(t, TypeDecided, nm.Msg.Type)
+	data, ok := nm.Msg.Data.([]*ParticipantsAPI)
 	require.True(t, ok, "response data should be participants slice")
 	require.Len(t, data, 1)
 	require.Equal(t, uint64(3), uint64(data[0].Slot))
@@ -232,8 +223,8 @@ func TestWSQuery_Decided_RangeMultipleSlots(t *testing.T) {
 
 	nm := h.QueryDecided(10, 12, hex.EncodeToString(pk[:]), spectypes.BNRoleProposer.String())
 
-	require.Equal(t, api.TypeDecided, nm.Msg.Type)
-	data, ok := nm.Msg.Data.([]*api.ParticipantsAPI)
+	require.Equal(t, TypeDecided, nm.Msg.Type)
+	data, ok := nm.Msg.Data.([]*ParticipantsAPI)
 	require.True(t, ok, "response data should be participants slice")
 	require.Len(t, data, 3)
 	require.Equal(t, uint64(10), uint64(data[0].Slot))
@@ -244,7 +235,7 @@ func TestWSQuery_Decided_RangeMultipleSlots(t *testing.T) {
 func TestWSQuery_InvalidPubKeyHex_ReturnsTypeError(t *testing.T) {
 	h := newWSQueryHarness(t)
 	nm := h.QueryDecided(1, 1, "zz-not-hex", spectypes.BNRoleProposer.String())
-	require.Equal(t, api.TypeError, nm.Msg.Type)
+	require.Equal(t, TypeError, nm.Msg.Type)
 	errs, ok := nm.Msg.Data.([]string)
 	require.True(t, ok, "expected []string, got %#v", nm.Msg.Data)
 	require.NotEmpty(t, errs)
@@ -260,7 +251,7 @@ func TestWSQuery_ValidatorNotFound_ReturnsTypeError(t *testing.T) {
 	h.ExpectValidator(pk, 0, false)
 
 	nm := h.QueryDecided(1, 1, "abcd", spectypes.BNRoleProposer.String())
-	require.Equal(t, api.TypeError, nm.Msg.Type)
+	require.Equal(t, TypeError, nm.Msg.Type)
 	errs, ok := nm.Msg.Data.([]string)
 	require.True(t, ok, "expected []string, got %#v", nm.Msg.Data)
 	require.Len(t, errs, 1)
@@ -276,7 +267,7 @@ func TestWSQuery_InvalidRole_ReturnsTypeError(t *testing.T) {
 
 	nm := h.QueryDecided(1, 1, hex.EncodeToString(pk[:]), "NOT_A_ROLE")
 
-	require.Equal(t, api.TypeError, nm.Msg.Type)
+	require.Equal(t, TypeError, nm.Msg.Type)
 	errs, ok := nm.Msg.Data.([]string)
 	require.True(t, ok, "expected []string, got %#v", nm.Msg.Data)
 	require.Len(t, errs, 1)
@@ -292,7 +283,7 @@ func TestWSQuery_TraceStoreError_NoEntries_ReturnsInternalError(t *testing.T) {
 	h.store.SetGetValidatorDutyError(phase0.Slot(1), spectypes.BNRoleProposer, idx, errors.New("boom"))
 
 	nm := h.QueryDecided(1, 1, hex.EncodeToString(pk[:]), spectypes.BNRoleProposer.String())
-	require.Equal(t, api.TypeError, nm.Msg.Type)
+	require.Equal(t, TypeError, nm.Msg.Type)
 	errs, ok := nm.Msg.Data.([]string)
 	require.True(t, ok, "expected []string, got %#v", nm.Msg.Data)
 	require.NotEmpty(t, errs)
@@ -309,7 +300,7 @@ func TestWSQuery_NoMessagesOnNotFound(t *testing.T) {
 
 	nm := h.QueryDecided(1, 1, hex.EncodeToString(pk[:]), spectypes.BNRoleProposer.String())
 
-	require.Equal(t, api.TypeDecided, nm.Msg.Type)
+	require.Equal(t, TypeDecided, nm.Msg.Type)
 	errs, ok := nm.Msg.Data.([]string)
 	require.True(t, ok, "expected []string, got %#v", nm.Msg.Data)
 	require.Len(t, errs, 1)
@@ -326,7 +317,7 @@ func TestWSQuery_Decided_Attester_NoMessagesWhenCommitteeNotFound(t *testing.T) 
 	// No committee duty link and no committee duty saved -> treated as "no messages".
 	nm := h.QueryDecided(3, 3, hex.EncodeToString(pk[:]), spectypes.BNRoleAttester.String())
 
-	require.Equal(t, api.TypeDecided, nm.Msg.Type)
+	require.Equal(t, TypeDecided, nm.Msg.Type)
 	errs, ok := nm.Msg.Data.([]string)
 	require.True(t, ok, "expected []string, got %#v", nm.Msg.Data)
 	require.Len(t, errs, 1)
@@ -344,8 +335,8 @@ func TestWSQuery_ErrorOnOneSlotButEntriesOnAnother_StillReturnsEntries(t *testin
 
 	nm := h.QueryDecided(10, 11, hex.EncodeToString(pk[:]), spectypes.BNRoleProposer.String())
 
-	require.Equal(t, api.TypeDecided, nm.Msg.Type)
-	data, ok := nm.Msg.Data.([]*api.ParticipantsAPI)
+	require.Equal(t, TypeDecided, nm.Msg.Type)
+	data, ok := nm.Msg.Data.([]*ParticipantsAPI)
 	require.True(t, ok, "response data should be participants slice")
 	require.Len(t, data, 1)
 	require.Equal(t, uint64(10), uint64(data[0].Slot))
@@ -354,10 +345,10 @@ func TestWSQuery_ErrorOnOneSlotButEntriesOnAnother_StillReturnsEntries(t *testin
 func TestWSQuery_NetworkMessageParseError_GoesThroughErrorHandler(t *testing.T) {
 	h := newWSQueryHarness(t)
 
-	nm := &api.NetworkMessage{Err: errors.New("parsefail")}
-	h.node.handleQueryRequests(nm)
+	nm := &NetworkMessage{Err: errors.New("parsefail")}
+	h.handler.HandleQueryRequests(nil, h.exporterRead, h.validatorMock, networkconfig.TestNetwork.DomainType, nm)
 
-	require.Equal(t, api.TypeError, nm.Msg.Type)
+	require.Equal(t, TypeError, nm.Msg.Type)
 	errs, ok := nm.Msg.Data.([]string)
 	require.True(t, ok, "expected []string, got %#v", nm.Msg.Data)
 	require.Len(t, errs, 2)
@@ -368,13 +359,13 @@ func TestWSQuery_NetworkMessageParseError_GoesThroughErrorHandler(t *testing.T) 
 func TestWSQuery_UnknownMessageType_ReturnsBadRequest(t *testing.T) {
 	h := newWSQueryHarness(t)
 
-	nm := &api.NetworkMessage{Msg: api.Message{
+	nm := &NetworkMessage{Msg: Message{
 		Type:   "banana",
-		Filter: api.MessageFilter{From: 1, To: 1},
+		Filter: MessageFilter{From: 1, To: 1},
 	}}
-	h.node.handleQueryRequests(nm)
+	h.handler.HandleQueryRequests(nil, h.exporterRead, h.validatorMock, networkconfig.TestNetwork.DomainType, nm)
 
-	require.Equal(t, api.TypeError, nm.Msg.Type)
+	require.Equal(t, TypeError, nm.Msg.Type)
 	errs, ok := nm.Msg.Data.([]string)
 	require.True(t, ok, "expected []string, got %#v", nm.Msg.Data)
 	require.Len(t, errs, 1)
@@ -390,7 +381,7 @@ func TestWSQuery_FromGreaterThanTo_ReturnsNoMessages(t *testing.T) {
 
 	nm := h.QueryDecided(11, 10, hex.EncodeToString(pk[:]), spectypes.BNRoleProposer.String())
 
-	require.Equal(t, api.TypeDecided, nm.Msg.Type)
+	require.Equal(t, TypeDecided, nm.Msg.Type)
 	errs, ok := nm.Msg.Data.([]string)
 	require.True(t, ok, "expected []string, got %#v", nm.Msg.Data)
 	require.Len(t, errs, 1)
